removed clients directory and interfaces
