+++ rtvi-client-js/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./actions";
-export * from "./clients";
+export * from "./client";
 export * from "./errors";
 export * from "./events";
 export * from "./helpers";
